Improvements
Improvements to code, including separating classes, in response to code review feedback
