+++ GameProject/Assets/TurnBasedAssets/Scripts/Dialogue/Message.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fcb92c4f8ba34eb8847485882b3db7eb
+timeCreated: 1563469433
