+++ GameProject/Assets/TurnBasedAssets/Scripts/Dialogue/NpcMoods.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace TurnBasedAssets.Scripts.Dialogue
+{
+    [Serializable]
+    public class NpcMoods
+    {
+        [SerializeField] private List<Image> _npcMoodImages;
+        public List<Image> NpcMoodImages => _npcMoodImages;
+    }
+}
