+++ GameProject/Assets/TurnBasedAssets/Scripts/Dialogue/NpcMoods.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 57b8714d83294ebfbc68d411b88225bc
+timeCreated: 1563469486
