+++ GameProject/Assets/TurnBasedAssets/Scripts/Dialogue/Response.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5172f115cfa54a6f960455a97cd256fc
+timeCreated: 1563469451
